Move files about a bit so that an interface can be written for the factory.
